Add Item Rosource

+++ src/main/java/com/example/demo/model/ui/ItemDto.java
@@ -4,12 +4,14 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
 public class ItemDto {
     private Long id;
     private String name;
-    private Float unitPrice;
+    private BigDecimal unitPrice;
     private Integer quantity;
 }
